backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
duckdb>=0.9.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

